backend/requirements.txt

fastapi
uvicorn
python-multipart
requests
pydantic
openai
python-dotenv
